Reestructuración de aplicación requirements y APIs de requisitos
