+++ requirements/models.py
@@ -5,6 +5,7 @@ from customers.models import Customer_type
 # Create your models here.
 class Requirement(models.Model):
     requirement_name = models.CharField(max_length=150)
+    is_active = models.BooleanField(default=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
